+++ cli/src/commands/bookmark/create.rs
@@ -81,9 +81,27 @@ pub fn cmd_bookmark_create(
     }
 
     let mut tx = workspace_command.start_transaction();
+    let remote_settings = tx.settings().remote_settings()?;
+    let readonly_repo = tx.base_repo().clone();
     for name in bookmark_names {
         tx.repo_mut()
             .set_local_bookmark_target(name, RefTarget::normal(target_commit.id().clone()));
+        for (remote_name, settings) in &remote_settings {
+            if !settings.auto_track_bookmarks.is_match(name.as_str()) {
+                continue;
+            }
+            let Some(view) = readonly_repo.view().get_remote_view(remote_name) else {
+                continue;
+            };
+            let symbol = name.to_remote_symbol(remote_name);
+            if view.bookmarks.contains_key(name) {
+                writeln!(
+                    ui.warning_default(),
+                    "Auto-tracking bookmark that exists on the remote: {symbol}"
+                )?;
+            }
+            tx.repo_mut().track_remote_bookmark(symbol)?;
+        }
     }
 
     if let Some(mut formatter) = ui.status_formatter() {