+++ cli/tests/test_git_fetch.rs
@@ -2047,3 +2047,49 @@ fn test_git_fetch_tracked_multiple_remotes() {
     [EOF]
     ");
 }
+
+#[test]
+fn test_git_fetch_auto_track_bookmarks() {
+    let test_env = TestEnvironment::default();
+    let root_dir = test_env.work_dir("");
+    test_env.add_config(
+        "
+        [remotes.origin]
+        auto-track-bookmarks = 'glob:mine/*'
+        ",
+    );
+
+    root_dir
+        .run_jj(["git", "init", "--colocate", "origin"])
+        .success();
+    let origin_dir = test_env.work_dir("origin");
+    origin_dir.run_jj(["b", "c", "mine/foo"]).success();
+    origin_dir.run_jj(["b", "c", "not-mine/foo"]).success();
+    origin_dir.run_jj(["commit", "-mfoo"]).success();
+    let output = origin_dir.run_jj(["show", "@-"]);
+    insta::assert_snapshot!(output, @r"
+    Commit ID: d7828da83253475bf10c2ae6bd3f0f84bf4604c1
+    Change ID: qpvuntsmwlqtpsluzzsnyyzlmlwvmlnu
+    Bookmarks: mine/foo not-mine/foo mine/foo@git not-mine/foo@git
+    Author   : Test User <test.user@example.com> (2001-02-03 08:05:10)
+    Committer: Test User <test.user@example.com> (2001-02-03 08:05:10)
+
+        foo
+
+    [EOF]
+    ");
+
+    root_dir.run_jj(["git", "init", "repo"]).success();
+    let repo_dir = test_env.work_dir("repo");
+    repo_dir
+        .run_jj(["git", "remote", "add", "origin", "../origin/.git"])
+        .success();
+
+    let output = repo_dir.run_jj(["git", "fetch"]);
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    bookmark: mine/foo@origin     [new] tracked
+    bookmark: not-mine/foo@origin [new] untracked
+    [EOF]
+    ");
+}